+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_infos/config.hpp
@@ -1,0 +1,11 @@
+class cfgWeapons
+{
+	//Include weapons
+	#include "weapons\config.hpp"
+};
+
+class cfgVehicles
+{
+	//Include vehicles
+	#include "vehicles\config.hpp"
+};
