+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_infos/vehicles/armour_standard/odst/config.hpp
@@ -1,0 +1,100 @@
+class 34thPRC_ArmourStandard_ODST_M56S_Urban
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban";
+	camo = "urban";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Arctic
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban";
+	camo = "arctic";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Arid
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban";
+	camo = "arid";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Tropic
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban";
+	camo = "tropic";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Woodland
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban";
+	camo = "woodland";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Urban_Guncase
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban_Guncase";
+	camo = "urban";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Arctic_Guncase
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban_Guncase";
+	camo = "arctic";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Arid_Guncase
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban_Guncase";
+	camo = "arid";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Tropic_Guncase
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban_Guncase";
+	camo = "tropic";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Woodland_Guncase
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban_Guncase";
+	camo = "woodland";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Urban_Medic
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban_Medic";
+	camo = "urban";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Arctic_Medic
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban_Medic";
+	camo = "arctic";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Arid_Medic
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban_Medic";
+	camo = "arid";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_tropic_Medic
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban_Medic";
+	camo = "tropic";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Woodland_Medic
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban_Medic";
+	camo = "woodland";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Urban_Radio
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban_Radio";
+	camo = "urban";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Arctic_Radio
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban_Radio";
+	camo = "arctic";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Arid_Radio
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban_Radio";
+	camo = "arid";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Tropic_Radio
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban_Radio";
+	camo = "tropic";
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Woodland_Radio
+{
+	model = "34thPRC_ArmourStandard_ODST_M56S_Urban_Radio";
+	camo = "woodland";
+};
