+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_infos/weapons/armour_custom/odst/yandere/config.hpp
@@ -1,0 +1,383 @@
+//Yandere Custom CH252D Helmets
+class 34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="urban";
+    visor="custom";
+};
+class 34thPRC_ArmourCustom_ODST_CH252D_Yandere_SilverVisor_Urban
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="urban";
+    visor="silver";
+};
+class 34thPRC_ArmourCustom_ODST_CH252D_Yandere_RedVisor_Urban
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="urban";
+    visor="red";
+};
+class 34thPRC_ArmourCustom_ODST_CH252D_Yandere_GoldVisor_Urban
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="urban";
+    visor="gold";
+};
+class 34thPRC_ArmourCustom_Yandere_CH252D_DP_Urban
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="urban";
+    visor="dp";
+};
+class 34thPRC_CustomArmour_ODST_CH2552D_Yandere_CustomVisor_Arid
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="arid";
+    visor="custom";
+};
+class 34thPRC_CustomArmour_ODST_CH252D_Yandere_SilverVisor_Arid
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="arid";
+    visor="silver";
+};
+class 34thPRC_CustomArmour_Yandere_ODST_CH252D_RedVisor_Aird
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="arid";
+    visor="red";
+};
+class 34thPRC_ArmourCustom_ODST_CH252D_Yandere_GoldVisor_Arid
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="arid";
+    visor="gold";
+};
+class 34thPRC_ArmourCustom_Yandere_CH252D_DP_Arid
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="arid";
+    visor="dp";
+};
+class 34thPRC_CustomArmour_ODST_CH2552D_Yandere_CustomVisor_Arctic
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="arctic";
+    visor="custom";
+};
+class 34thPRC_CustomArmour_ODST_CH252D_Yandere_SilverVisor_Arctic
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="arctic";
+    visor="silver";
+};
+class 34thPRC_CustomArmour_Yandere_ODST_CH252D_RedVisor_Aird
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="arctic";
+    visor="red";
+};
+class 34thPRC_ArmourCustom_ODST_CH252D_Yandere_GoldVisor_Arctic
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="arctic";
+    visor="gold";
+};
+class 34thPRC_ArmourCustom_Yandere_CH252D_DP_Arctic
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="arctic";
+    visor="dp";
+};
+class 34thPRC_CustomArmour_ODST_CH2552D_Yandere_CustomVisor_Tropic
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="tropic";
+    visor="custom";
+};
+class 34thPRC_CustomArmour_ODST_CH252D_Yandere_SilverVisor_Tropic
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="tropic";
+    visor="silver";
+};
+class 34thPRC_CustomArmour_Yandere_ODST_CH252D_RedVisor_Aird
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="tropic";
+    visor="red";
+};
+class 34thPRC_ArmourCustom_ODST_CH252D_Yandere_GoldVisor_Tropic
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="tropic";
+    visor="gold";
+};
+class 34thPRC_ArmourCustom_Yandere_CH252D_DP_Tropic
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="tropic";
+    visor="dp";
+};
+class 34thPRC_CustomArmour_ODST_CH2552D_Yandere_CustomVisor_Woodland
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="woodland";
+    visor="custom";
+};
+class 34thPRC_CustomArmour_ODST_CH252D_Yandere_SilverVisor_Woodland
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="woodland";
+    visor="silver";
+};
+class 34thPRC_CustomArmour_Yandere_ODST_CH252D_RedVisor_Aird
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="woodland";
+    visor="red";
+};
+class 34thPRC_ArmourCustom_ODST_CH252D_Yandere_GoldVisor_Woodland
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="woodland";
+    visor="gold";
+};
+class 34thPRC_ArmourCustom_Yandere_CH252D_DP_Woodland
+{
+    model="34thPRC_ArmourCustom_ODST_CH2552D_Yandere_CustomVisor_Urban";
+    player="yandere";
+    camo="woodland";
+    visor="dp";
+};
+
+//Yandere Custom M52
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_Urban
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="urban";
+    variant="a";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_B_Urban
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="urban";
+    variant="b";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_C_Urban
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="urban";
+    variant="c";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_Marksman_Urban
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    camo="urban";
+    variant="marksman";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_CQB_Urban
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="urban";
+    variant="cqb";
+};
+class 34thPRC_ArmourCustom_ODST_Yanndere_M52_Medic_Urban
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="urban";
+    variant="medic";
+};
+class 34thPRC_CustomArmour_Yandere_ODST_M52_Arid
+{
+    model="34thPRC_CustomArmour_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="arid";
+    variant="a";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_B_Arid
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="arid";
+    variant="b";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_C_Arid
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="arid";
+    variant="c";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_Marksman_Arid
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    camo="aird";
+    variant="marksman";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_CQB_Arid
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="arid";
+    variant="cqb";
+};
+class 34thPRC_ArmourCustom_ODST_Yanndere_M52_Medic_Arid
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="arid";
+    variant="medic";
+};
+class 34thPRC_CustomArmour_Yandere_ODST_M52_Arctic
+{
+    model="34thPRC_CustomArmour_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="arctic";
+    variant="a";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_B_Arctic
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="arctic";
+    variant="b";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_C_Arctic
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="arctic";
+    variant="c";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_Marksman_Arctic
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    camo="arctic";
+    variant="marksman";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_CQB_Arctic
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="arctic";
+    variant="cqb";
+};
+class 34thPRC_ArmourCustom_ODST_Yanndere_M52_Medic_Arctic
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="arctic";
+    variant="medic";
+};
+class 34thPRC_CustomArmour_Yandere_ODST_M52_Tropic
+{
+    model="34thPRC_CustomArmour_ODST_M52_Yandere_Tropic";
+    player="yandere";
+    camo="tropic";
+    variant="a";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_B_Tropic
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="tropic";
+    variant="b";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_C_Tropic
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="tropic";
+    variant="c";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_Marksman_Tropic
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    camo="tropic";
+    variant="marksman";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_CQB_Tropic
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="tropic";
+    variant="cqb";
+};
+class 34thPRC_ArmourCustom_ODST_Yanndere_M52_Medic_Tropic
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="tropic";
+    variant="medic";
+};
+class 34thPRC_CustomArmour_Yandere_ODST_M52_Woodland
+{
+    model="34thPRC_CustomArmour_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="woodland";
+    variant="a";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_B_Woodland
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="woodland";
+    variant="b";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_C_Woodland
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="woodland";
+    variant="c";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_Marksman_Woodland
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    camo="woodland";
+    variant="marksman";
+};
+class 34thPRC_ArmourCustom_Yandere_ODST_M52_CQB_Woodland
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="woodland";
+    variant="cqb";
+};
+class 34thPRC_ArmourCustom_ODST_Yanndere_M52_Medic_Woodland
+{
+    model="34thPRC_ArmourCustom_ODST_M52_Yandere_Urban";
+    player="yandere";
+    camo="woodland";
+    variant="medic";
+};
