+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_infos/weapons/armour_standard/config.hpp
@@ -1,0 +1,3 @@
+#include "marines2\config.hpp"
+#include "odst\config.hpp"
+#include "odst2\config.hpp"
