+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_infos/weapons/armour_standard/odst/config.hpp
@@ -1,0 +1,172 @@
+// MA ODST Vests
+class 34thPRC_ArmourStandard_ODST_M56R_Urban
+{
+	model = "34thPRC_ArmourStandard_ODST_M56R_Urban";
+	camo = "urban";
+	medic = "standard";
+	shoulders = "standard";
+};
+class 34thPRC_ArmourStandard_ODST_M56R_Arid
+{
+	model = "34thPRC_ArmourStandard_ODST_M56R_Urban";
+	camo = "arid";
+	medic = "standard";
+	shoulders = "standard";
+}
+class 34thPRC_ArmourStandard_ODST_M56R_Arctic
+{
+	model = "34thPRC_ArmourStandard_ODST_M56R_Urban";
+	camo = "arctic";
+	medic = "standard";
+	shoulders = "standard";
+}
+class 34thPRC_ArmourStandard_ODST_M56R_Woodland
+{
+	model = "34thPRC_ArmourStandard_ODST_M56R_Urban";
+	camo = "woodland";
+	medic = "standard";
+	shoulders = "standard";
+}
+class 34thPRC_ArmourStandard_ODST_M56R_Tropic
+{
+	model = "34thPRC_ArmourStandard_ODST_M56R_Urban";
+	camo = "tropic";
+	medic = "standard";
+	shoulders = "standard";
+}
+class 34thPRC_ArmourStandard_ODST_M56R_Urban_Medic
+{
+	model = "34thPRC_ArmourStandard_ODST_M56R_Urban";
+	camo = "urban";
+	medic = "isMedic";
+	shoulders = "standard";
+}
+class 34thPRC_ArmourStandard_ODST_M56R_Arctic_Medic
+{
+	model = "34thPRC_ArmourStandard_ODST_M56R_Urban";
+	camo = "arctic";
+	medic = "isMedic";
+	shoulders = "standard";
+}
+class 34thPRC_ArmourStandard_ODST_M56R_Arid_Medic
+{
+	model = "34thPRC_ArmourStandard_ODST_M56R_Urban";
+	camo = "arid";
+	medic = "isMedic";
+	shoulders = "standard";
+}
+class 34thPRC_ArmourStandard_ODST_M56R_Tropic_Medic
+{
+	model = "34thPRC_ArmourStandard_ODST_M56R_Urban";
+	camo = "tropic";
+	medic = "isMedic";
+	shoulders = "standard";
+}
+class 34thPRC_ArmourStandard_ODST_M56R_Woodland_Medic
+{
+	model = "34thPRC_ArmourStandard_ODST_M56R_Urban";
+	camo = "woodland";
+	medic = "isMedic";
+	shoulders = "standard";
+}
+class 34thPRC_ArmourStandard_ODST_M56R_Urban_CQB
+{
+	model = "34thPRC_ArmourStandard_ODST_M56R_Urban";
+	camo = "urban";
+	medic = "standard";
+	shoulders = "cqb";
+}
+
+// MA ODST Uniforms
+class 34thPRC_ArmourStandard_ODST_Uniform_UrbanRed
+{
+	model = "34thPRC_ArmourStandard_ODST_Uniform_UrbanRed";
+	camo = "urbanRed";
+};
+class 34thPRC_ArmourStandard_ODST_Uniform_UrbanBlue
+{
+	model = "34thPRC_ArmourStandard_ODST_Uniform_UrbanRed";
+	camo = "urbanBlue";
+};
+class 34thPRC_ArmourStandard_ODST_Uniform_Arctic
+{
+	model = "34thPRC_ArmourStandard_ODST_Uniform_UrbanRed";
+	camo = "arctic";
+};
+class 34thPRC_ArmourStandard_ODST_Uniform_Arid
+{
+	model = "34thPRC_ArmourStandard_ODST_Uniform_UrbanRed";
+	camo = "arid";
+};
+class 34thPRC_ArmourStandard_ODST_Uniform_Tropic
+{
+	model = "34thPRC_ArmourStandard_ODST_Uniform_UrbanRed";
+	camo = "tropic";
+};
+class 34thPRC_ArmourStandard_ODST_Uniform_Woodland
+{
+	model = "34thPRC_ArmourStandard_ODST_Uniform_UrbanRed";
+	camo = "woodland";
+};
+
+// MA ODST Helmets
+class 34thPRC_ArmourStandard_ODST_M56SR_Urban_WhiteVisor
+{
+	model = "34thPRC_ArmourStandard_ODST_M56SR_Urban_WhiteVisor";
+	camo = "urban";
+	visor = "white";
+};
+class 34thPRC_ArmourStandard_ODST_M56SR_Urban_MarooonVisor
+{
+	model = "34thPRC_ArmourStandard_ODST_M56SR_Urban_WhiteVisor";
+	camo = "urban";
+	visor = "maroon";
+};
+class 34thPRC_ArmourStandard_ODST_M56SR_Arctic_WhiteVisor
+{
+	model = "34thPRC_ArmourStandard_ODST_M56SR_Urban_WhiteVisor";
+	camo = "arctic";
+	visor = "white";
+};
+class 34thPRC_ArmourStandard_ODST_M56SR_Arctic_MarooonVisor
+{
+	model = "34thPRC_ArmourStandard_ODST_M56SR_Urban_WhiteVisor";
+	camo = "arctic";
+	visor = "maroon";
+};
+class 34thPRC_ArmourStandard_ODST_M56SR_Arid_WhiteVisor
+{
+	model = "34thPRC_ArmourStandard_ODST_M56SR_Urban_WhiteVisor";
+	camo = "arid";
+	visor = "white";
+};
+class 34thPRC_ArmourStandard_ODST_M56SR_Arid_MarooonVisor
+{
+	model = "34thPRC_ArmourStandard_ODST_M56SR_Urban_WhiteVisor";
+	camo = "arid";
+	visor = "maroon";
+};
+class 34thPRC_ArmourStandard_ODST_M56SR_Tropic_WhiteVisor
+{
+	model = "34thPRC_ArmourStandard_ODST_M56SR_Urban_WhiteVisor";
+	camo = "tropic";
+	visor = "white";
+};
+class 34thPRC_ArmourStandard_ODST_M56SR_Tropic_MarooonVisor
+{
+	model = "34thPRC_ArmourStandard_ODST_M56SR_Urban_WhiteVisor";
+	camo = "tropic";
+	visor = "maroon";
+};
+class 34thPRC_ArmourStandard_ODST_M56SR_Woodland_WhiteVisor
+{
+	model = "34thPRC_ArmourStandard_ODST_M56SR_Urban_WhiteVisor";
+	camo = "woodland";
+	visor = "white";
+};
+class 34thPRC_ArmourStandard_ODST_M56SR_Woodland_MarooonVisor
+{
+	model = "34thPRC_ArmourStandard_ODST_M56SR_Urban_WhiteVisor";
+	camo = "woodland";
+	visor = "maroon";
+};
