+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_infos/weapons/config.hpp
@@ -1,0 +1,1 @@
+#include "armour_standard\config.hpp"
