+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/vehicles/armour_standard/odst/config.hpp
@@ -1,0 +1,128 @@
+class 34thPRC_ArmourStandard_ODST_M56S_Urban
+{
+	label = "[34thPRC] M56S ODST Rucksack";
+	author = "34th PRC Modding Team";
+	options[] = {"camo"};
+	class camo
+	{
+		alwaysSelectable = 1;
+		changeingame = 0;
+		values[] = {"urban", "arctic", "arid", "tropic", "woodland"};
+		class urban
+		{
+			label = "Urban";
+		};
+		class arctic
+		{
+			label = "Arctic";
+		};
+		class arid
+		{
+			label = "Arid";
+		};
+		class tropic
+		{
+			label = "Tropic";
+		};
+		class woodland
+		{
+			label = "Woodland";
+		};
+	};
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Urban_Guncase
+{
+	label = "[34thPRC] M56S ODST Guncase";
+	author = "34th PRC Modding Team";
+	options[] = {"camo"};
+	class camo
+	{
+		alwaysSelectable = 1;
+		changeingame = 0;
+		values[] = {"urban", "arctic", "arid", "tropic", "woodland"};
+		class urban
+		{
+			label = "Urban";
+		};
+		class arctic
+		{
+			label = "Arctic";
+		};
+		class arid
+		{
+			label = "Arid";
+		};
+		class tropic
+		{
+			label = "Tropic";
+		};
+		class woodland
+		{
+			label = "Woodland";
+		};
+	};
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Urban_Medic
+{
+	label = "[34thPRC] M56S ODST Medic Rucksack";
+	author = "34th PRC Modding Team";
+	options[] = {"camo"};
+	class camo
+	{
+		alwaysSelectable = 1;
+		changeingame = 0;
+		values[] = {"urban", "arctic", "arid", "tropic", "woodland"};
+		class urban
+		{
+			label = "Urban";
+		};
+		class arctic
+		{
+			label = "Arctic";
+		};
+		class arid
+		{
+			label = "Arid";
+		};
+		class tropic
+		{
+			label = "Tropic";
+		};
+		class woodland
+		{
+			label = "Woodland";
+		};
+	};
+};
+class 34thPRC_ArmourStandard_ODST_M56S_Urban_Radio
+{
+	label = "[34thPRC] M56S ODST Radio Rucksack";
+	author = "34th PRC Modding Team";
+	options[] = {"camo"};
+	class camo
+	{
+		alwaysSelectable = 1;
+		changeingame = 0;
+		values[] = {"urban", "arctic", "arid", "tropic", "woodland"};
+		class urban
+		{
+			label = "Urban";
+		};
+		class arctic
+		{
+			label = "Arctic";
+		};
+		class arid
+		{
+			label = "Arid";
+		};
+		class tropic
+		{
+			label = "Tropic";
+		};
+		class woodland
+		{
+			label = "Woodland";
+		};
+	};
+};
