+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/weapons/armour_custom/config.hpp
@@ -1,0 +1,2 @@
+//#include "marines2\config.hpp"
+#include "odst\config.hpp"
