+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/weapons/armour_custom/odst/config.hpp
@@ -1,0 +1,1 @@
+#include "group1\config.hpp"
