+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/weapons/armour_standard/odst2/config.hpp
@@ -1,0 +1,120 @@
+//Ma ODST Armor New
+class 34thPRC_ArmourStandard_ODST_M52_Urban
+{
+	label = "[34thPRC] ODST Armors";
+	author = "34th PRC Modding Team, Over Yandere";
+	options[] = {"camo", "variant"};
+	class camo
+	{
+		alwaysSelectable = 1;
+		changeingame = 0;
+		values[] = {"urban", "arctic", "arid", "tropic", "woodland"};
+		class urban
+		{
+			label = "Urban";
+		};
+		class arctic
+		{
+			label = "Arctic";
+		};
+		class arid
+		{
+			label = "Arid";
+		};
+		class tropic
+		{
+			label = "Tropic";
+		};
+		class woodland
+		{
+			label = "Woodland";
+		};
+	};
+	class variant
+	{
+		alwaysSelectable = 1;
+		changeingame = 0;
+		values[] = {"a", "b", "c", "cqb", "marksman", "medic"};
+		class a
+		{
+			label = "M52-A";
+		};
+		class b
+		{
+			label = "M52-B";
+		};
+		class c
+		{
+			label = "M52-C";
+		};
+		class cqb
+		{
+			label = "CQB";
+		};
+		class marksman
+		{
+			label = "Marksman";
+		};
+		class medic
+		{
+			label = "Medic";
+		};
+	};
+};
+
+//MA ODST Helmet New
+class 34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban
+{
+	label = "[34thPRC] ODST CH252D Helmets";
+	author = "34th PRC Modding Team, Over Yandere";
+	options[] = {"camo", "visor"};
+	class camo
+	{
+		alwaysSelectable = 1;
+		changeingame = 0;
+		values[] = {"urban", "arctic", "arid", "tropic", "woodland"};
+		class urban
+		{
+			label = "Urban";
+		};
+		class arctic
+		{
+			label = "Arctic";
+		};
+		class arid
+		{
+			label = "Arid";
+		};
+		class tropic
+		{
+			label = "Tropic";
+		};
+		class woodland
+		{
+			label = "Woodland";
+		};
+	};
+	class visor
+	{
+		alwaysSelectable = 1;
+		changeingame = 0;
+		values[] = {"silver", "red", "gold", "dp"};
+		class silver
+		{
+			label = "Silver";
+		};
+		class maroon
+		{
+			label = "Red";
+		};
+		class gold
+		{
+			label = "Gold";
+		};
+		class dp
+		{
+			label = "Depolarized";
+		};
+	};
+};
+
