+++ src/Addons/34thPRC_ArmourCustom/config.cpp
@@ -12,9 +12,19 @@ class CfgPatches
 };
 class CfgVehicles
 {
-	#include "data\config_vehicles.hpp"
+	//#include "data\config_vehicles.hpp"
 };
 class CfgWeapons
 {
+	//Defines for custom ODST vest & helmet
+	class 34thPRC_ArmourStandard_ODST_CH252D_Urban_SilverVisor; //34thPRC_ArmourStandard
+	class 34thPRC_ArmourStandard_ODST_CH252D_Urban_DP; //34thPRC_ArmourStandard
+	class 34thPRC_ArmourStandard_ODST_M52_Urban; //34thPRC_ArmourStandard //Variant A / default
+	class 34thPRC_ArmourStandard_ODST_M52_Marksman_Urban; //34thPRC_ArmourStandard //Marksman shoulders and bobs
+	class 34thPRC_ArmourStandard_ODST_M52_CQB_Urban; //34thPRC_ArmourStandard //CQB Shoulders and bobs
+	class 34thPRC_ArmourStandard_ODST_M52_B_Urban; //34thPRC_ArmourStandard //Variant B
+	class 34thPRC_ArmourStandard_ODST_M52_C_Urban; //34thPRC_ArmourStandard //Variant C
+
+	//includes
 	#include "data\config_weapons.hpp"
 };
