+++ src/Addons/34thPRC_ArmourCustom/data/config_weapons.hpp
@@ -1,2 +1,2 @@
-#include "odst\config_weapons.hpp"
+#include "marines\config_weapons.hpp"
 #include "odst\config_weapons.hpp"
