--- src/Addons/34thPRC_ArmourStandard/data/odst2/34th_CH252D.rvmat
@@ -1,123 +1,0 @@
-class 34th_CH252D
-{
-//Type: 11 (Arma3/VBS2)
-//Arma1 flag= 1;
-mainLight=Sun;//(1)
-ul_FogMode=Fog;//(1)
-Ambient[]=      {1,1,1,1};
-Diffuse[]=      {1,1,1,1};
-Emmisive[]=     {0,0,0,1};
-ForcedDiffuse[]=        {0,0,0,0};
-Specular[]=     {0.2,0.2,0.2,1};
-//Specular2[]=  {0.2,0.2,0.2,1};
-SpecularPower=25;
-PixelShaderId=Super;//102
-VertexShaderId=Super;//23
-class TegGen0
-{
-        uvSource=None;
-};
-class TegGen1
-{
-        uvSource=Tex;
-        class uvTransform
-        {
-                aside[] =       {   1.0000,   0.0000,   0.0000};
-                up[] =  {   0.0000,   1.0000,   0.0000};
-                dir[] = {   0.0000,   0.0000,   0.0000};
-                pos[] = {   0.0000,   0.0000,   0.0000};
-        };
-};
-class TegGen2
-{
-        uvSource=Tex;
-        class uvTransform
-        {
-                aside[] =       {   1.0000,   1.0000,   0.0000};
-                up[] =  {   0.0000,   1.0000,   0.0000};
-                dir[] = {   0.0000,   0.0000,   0.0000};
-                pos[] = {   0.0000,   0.0000,   0.0000};
-        };
-};
-class TegGen3
-{
-        uvSource=Tex;
-        class uvTransform
-        {
-                aside[] =       {   1.0000,   0.0000,   0.0000};
-                up[] =  {   0.0000,   1.0000,   0.0000};
-                dir[] = {   0.0000,   0.0000,   0.0000};
-                pos[] = {   0.0000,   0.0000,   0.0000};
-        };
-};
-class TegGen4
-{
-        uvSource=Tex;
-        class uvTransform
-        {
-                aside[] =       {   1.0000,   0.0000,   0.0000};
-                up[] =  {   0.0000,   1.0000,   0.0000};
-                dir[] = {   0.0000,   0.0000,   1.0000};
-                pos[] = {   0.0000,   0.0000,   1.0000};
-        };
-};
-class TegGen5
-{
-        uvSource=Tex;
-        class uvTransform
-        {
-                aside[] =       {   1.0000,   0.0000,   0.0000};
-                up[] =  {   0.0000,   1.0000,   0.0000};
-                dir[] = {   0.0000,   0.0000,   0.0000};
-                pos[] = {   0.0000,   0.0000,   0.0000};
-        };
-};
-class TegGen6
-{
-        uvSource=Tex;
-};
-class TegGen7
-{
-        uvSource=Tex;
-};
-class stage1
-{
-        texture="34thPRC_ArmourStandard\data\odst2\34th_odst_helmet_NOHQ.paa";
-        Texgen=1;
-};
-class stage2
-{
-        texture="#(argb,8,8,3)color(0.5,0.5,0.5,0.5,dt)";
-        Texgen=2;
-};
-class stage3
-{
-        texture="#(argb,8,8,3)color(0,0,0,0,mc)";
-        Texgen=3;
-};
-class stage4
-{
-        texture="34thPRC_ArmourStandard\data\odst2\34th_ODST_Helmet_AS.paa";
-        Texgen=4;
-};
-class stage5
-{
-        texture="34thPRC_ArmourStandard\data\odst2\34th_odst_helmet_SMDI.paa";
-        Texgen=5;
-};
-class stage6
-{
-        texture="#(ai,64,64,1)fresnel(3,2)";
-        Texgen=6;
-};
-class stage7
-{
-        useWorldEnvMap=true;
-        texture="a3\data_f\env_land_ca.paa";
-        Texgen=7;
-};
-class stageTI
-{
-        texture="a3\data_f\default_vehicle_ti_ca.paa";
-};
-};
