--- src/Addons/34thPRC_ArmourStandard/data/odst2/34th_ODST_Vest_AS.paa
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:6d76240fe5dbb29a8de4144c3ef71ebf1bf47af33bdbb7feef9a97ad0b2dbae8
-size 1566490
