--- src/Addons/34thPRC_ArmourStandard/data/odst2/34th_ODST_Vest_SMDI.paa
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:1755d130e08620f7a44fd40c3487e28ec573b979b05d859347457b69b8d98e94
-size 1926981
